+++ src/client/deploy-commands.ts
@@ -4,7 +4,6 @@ import { Routes } from 'discord-api-types/v9';
 import config from '../config';
 import chalk from 'chalk';
 import * as commandModules from '../commands';
-import { Collection } from 'discord.js';
 
 type Command = {
   data: Pick<SlashCommandBuilder, 'toJSON'>;
@@ -15,7 +14,6 @@ const commands: Pick<SlashCommandBuilder, 'toJSON'>[] = [];
 for (const module of Object.values<Command>(commandModules)) {
   if (module.data && typeof module.data.toJSON === 'function') {
     commands.push(module.data);
-    console.log(module.data);
   }
 }
